Cleanup command runner
We had a very elaborate scheme for testing command execution,
involving calling back in to the test program, but we weren't really
using it, much simpler to just use a mock.

Also avoid hyphenated package naming, which isn't traditional
kubernetes / go, and confuses IDEs / static analysis tools.
