+++ pkg/cmdrunner/cmd_runner.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cmd_runner
+package cmdrunner
 
 import (
 	"os/exec"
