+++ pkg/kubeadm/BUILD.bazel
@@ -5,12 +5,12 @@ go_library(
     srcs = ["kubeadm.go"],
     importpath = "sigs.k8s.io/cluster-api/pkg/kubeadm",
     visibility = ["//visibility:public"],
-    deps = ["//pkg/cmd-runner:go_default_library"],
+    deps = ["//pkg/cmdrunner:go_default_library"],
 )
 
 go_test(
     name = "go_default_test",
     srcs = ["kubeadm_test.go"],
     embed = [":go_default_library"],
-    deps = ["//pkg/test-cmd-runner:go_default_library"],
+    deps = ["//pkg/testcmdrunner:go_default_library"],
 )
