--- pkg/test-cmd-runner/BUILD.bazel
@@ -1,14 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["test_cmd_runner.go"],
-    importpath = "sigs.k8s.io/cluster-api/pkg/test-cmd-runner",
-    visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["test_cmd_runner_test.go"],
-    embed = [":go_default_library"],
-)
